p3-docker-deploy/requirements.txt
bcrypt==4.0.1
cffi==1.16.0
cryptography==41.0.4
paramiko==3.3.1
pycparser==2.21
PyNaCl==1.5.0
PyYAML==6.0.1
scp==0.14.5
